Accounts

+++ main/Settings/Accounts/views.py
@@ -1,10 +1,11 @@
 from flask import Blueprint,jsonify,request
-from main.Settings.Accounts.models import BankAccounts
+from main.Settings.Accounts.models import BankAccounts,CategorySubcategory
 from main.Settings.Funds.models import MasterData
 from main.extensions import db
 
 settings_accounts=Blueprint('settings-accounts',__name__,url_prefix='/settings-accounts')
 
+#category
 @settings_accounts.route('/add-category',methods=["POST"])
 def addCategory():
     try:
@@ -52,6 +53,10 @@ def showAllCategories():
 def deleteCategory(id):
     try:
         entry=MasterData.query.get(id)
+        if not entry:
+            return jsonify({
+                "message":"category not exist."
+            })
         db.session.delete(entry)
         db.session.commit()
         return jsonify({
@@ -62,6 +67,82 @@ def deleteCategory(id):
             "error":str(e)
         })
 
+#Expense
+@settings_accounts.route('/add-expense',methods=['POST'])
+def addExpense():
+    try:
+        data=request.get_json()
+        TYPE='Expense'
+        CATEGORY=data.get('CATEGORY')
+        SUBCATEGORY=data.get('SUBCATEGORY')
+        expenses=CategorySubcategory.query.filter(CategorySubcategory.SUBCATEGORY==SUBCATEGORY)
+        subCategoryExist=False
+        for i in expenses:
+            subCategoryExist=True
+        if subCategoryExist:
+            return jsonify({
+                "message":"expense already exist"
+            })
+        entry=CategorySubcategory(TYPE=TYPE,
+                         CATEGORY=CATEGORY,
+                         SUBCATEGORY=SUBCATEGORY)
+        db.session.add(entry)
+        db.session.commit()
+        return jsonify({
+            "id":entry.id,
+            "TYPE":entry.TYPE,
+            "CATEGORY":entry.CATEGORY,
+            "SUBCATEGORY":entry.SUBCATEGORY
+            })
+    except Exception as e:
+        return jsonify({
+            "error":str(e)
+        })
+
+@settings_accounts.route('/show-all-expenses')
+def showAllExpenses():
+    try:
+        expenses=CategorySubcategory.query.filter(CategorySubcategory.TYPE=='Expense')
+        data=[]
+        for i in expenses:
+            id=i.id
+            type=i.TYPE
+            category=i.CATEGORY
+            subcategory=i.SUBCATEGORY
+            info={
+                "id":id,
+                "TYPE":type,
+                "CATEGORY":category,
+                "SUBCATEGORY":subcategory
+            }
+            data.append(info)
+        return jsonify({
+            "data":data
+        })
+    except Exception as e:
+        return jsonify({
+            "error":str(e)
+        })
+
+@settings_accounts.route('/delete-expense/<int:id>',methods=["DELETE"])
+def deleteExpense(id):
+    try:
+        entry=CategorySubcategory.query.get(id)
+        if not entry:
+            return jsonify({
+                "message":"expense not exist."
+            })
+        db.session.delete(entry)
+        db.session.commit()
+        return jsonify({
+            "messgae":"category deleted successfully."
+        })
+    except Exception as e:
+        return jsonify({
+            "error":str(e)
+        })
+
+#Bank Details
 @settings_accounts.route('/add-bank-details',methods=['POST'])
 def addBankDetails():
     try:
